Extract SI unit formatting and parsing to a dedicated value object, removing static methods from the `FileInformation` class
