+++ src/File/FileInformation.php
@@ -9,22 +9,15 @@ use Psr\Http\Message\UploadedFileInterface;
 
 use function assert;
 use function basename;
-use function ctype_digit;
 use function file_exists;
 use function filesize;
 use function finfo_open;
 use function is_array;
 use function is_int;
-use function is_numeric;
 use function is_readable;
 use function is_string;
-use function round;
-use function strtoupper;
-use function substr;
-use function trim;
 
 use const FILEINFO_MIME_TYPE;
-use const PHP_INT_MAX;
 
 /** @internal */
 final class FileInformation
@@ -37,25 +30,20 @@ final class FileInformation
         public readonly string $path,
         public readonly ?string $clientFileName,
         public readonly ?string $clientMediaType,
-        private int|null $size,
+        private Bytes|null $size,
     ) {
         $this->readable  = is_readable($this->path);
         $this->baseName  = basename($this->path);
         $this->mediaType = null;
     }
 
-    public function size(): int
+    public function size(): Bytes
     {
         if ($this->size === null) {
-            $this->size = filesize($this->path);
+            $this->size = Bytes::fromInteger(filesize($this->path));
         }
 
         return $this->size;
-    }
-
-    public function sizeAsSiUnit(): string
-    {
-        return self::bytesToSiUnit($this->size());
     }
 
     public function detectMimeType(): string
@@ -84,11 +72,13 @@ final class FileInformation
             $path = $value->getStream()->getMetadata('uri');
             assert(is_string($path));
 
+            $size = $value->getSize();
+
             return new self(
                 $path,
                 $value->getClientFilename(),
                 $value->getClientMediaType(),
-                $value->getSize(),
+                $size === null ? $size : Bytes::fromInteger($size),
             );
         }
 
@@ -110,6 +100,7 @@ final class FileInformation
         assert(is_string($clientName));
         assert(is_string($clientType));
         assert(is_int($size) || $size === null);
+        $size = $size === null ? $size : Bytes::fromInteger($size);
 
         return new self($path, $clientName, $clientType, $size);
     }
@@ -135,78 +126,5 @@ final class FileInformation
         }
 
         return false;
-    }
-
-    /**
-     * Format filesize in bytes to an SI Unit
-     */
-    public static function bytesToSiUnit(int $size): string
-    {
-        $sizes = ['B', 'kB', 'MB', 'GB', 'TB', 'PB', 'EB', 'ZB', 'YB'];
-        for ($i = 0; $size >= 1024 && $i < 9; $i++) {
-            $size /= 1024;
-        }
-
-        $suffix = $sizes[$i] ?? null;
-
-        assert(is_string($suffix));
-
-        return round($size, 2) . $suffix;
-    }
-
-    /**
-     * Convert an SI unit to bytes
-     */
-    public static function siUnitToBytes(string $size): int
-    {
-        if (ctype_digit($size)) {
-            return (int) $size;
-        }
-
-        $type = trim(substr($size, -2, 1));
-
-        $value = substr($size, 0, -1);
-        if (! is_numeric($value)) {
-            $value = trim(substr($value, 0, -1));
-        }
-
-        assert(is_numeric($value));
-
-        switch (strtoupper($type)) {
-            case 'Y':
-                //$value *= 1024 ** 8;
-                $value = PHP_INT_MAX;
-                break;
-            case 'Z':
-                //$value *= 1024 ** 7;
-                $value = PHP_INT_MAX;
-                break;
-            case 'E':
-                if ($value > 7) {
-                    $value = PHP_INT_MAX;
-                    break;
-                }
-                $value *= 1024 ** 6;
-                break;
-            case 'P':
-                $value *= 1024 ** 5;
-                break;
-            case 'T':
-                $value *= 1024 ** 4;
-                break;
-            case 'G':
-                $value *= 1024 ** 3;
-                break;
-            case 'M':
-                $value *= 1024 ** 2;
-                break;
-            case 'K':
-                $value *= 1024;
-                break;
-            default:
-                break;
-        }
-
-        return (int) $value;
     }
 }
