+++ src/File/FilesSize.php
@@ -73,11 +73,11 @@ final class FilesSize extends AbstractValidator
         }
 
         if (is_string($min)) {
-            $min = FileInformation::siUnitToBytes($min);
+            $min = Bytes::fromSiUnit($min)->bytes;
         }
 
         if (is_string($max)) {
-            $max = FileInformation::siUnitToBytes($max);
+            $max = Bytes::fromSiUnit($max)->bytes;
         }
 
         $this->min = $min !== null ? (int) $min : null;
@@ -94,10 +94,10 @@ final class FilesSize extends AbstractValidator
         );
 
         $this->minString = $this->min !== null && $this->useByteString
-            ? FileInformation::bytesToSiUnit($this->min)
+            ? Bytes::fromInteger($this->min)->toSiUnit()
             : (string) $this->min;
         $this->maxString = $this->max !== null && $this->useByteString
-            ? FileInformation::bytesToSiUnit($this->max)
+            ? Bytes::fromInteger($this->max)->toSiUnit()
             : (string) $this->max;
 
         parent::__construct($options);
@@ -143,11 +143,11 @@ final class FilesSize extends AbstractValidator
 
             $paths[] = $file->path;
 
-            $size += $file->size();
+            $size += $file->size()->bytes;
         }
 
         $this->size = $this->useByteString
-            ? FileInformation::bytesToSiUnit($size)
+            ? Bytes::fromInteger($size)->toSiUnit()
             : (string) $size;
 
         if ($this->min !== null && $size < $this->min) {
