+++ test/File/FilesSizeTest.php
@@ -69,6 +69,7 @@ final class FilesSizeTest extends TestCase
             'minimum: 0 byte; maximum: 2 megabytes; 2 MB'   => [['min' => 0, 'max' => '2 MB'],  true,  true,  true],
             'minimum: 0 byte; maximum: 2 megabytes; 2MB'    => [['min' => 0, 'max' => '2MB'],   true,  true,  true],
             'minimum: 0 byte; maximum: 2 megabytes; 2  MB'  => [['min' => 0, 'max' => '2  MB'], true,  true,  true],
+            'minimum: 1k; maximum: 2 megabytes; 2  MB'      => [['min' => '1k', 'max' => '2  MB'], true,  true,  true],
             // phpcs:enable
         ];
     }
